[refactor] [#21] SkillList Organism 이름 변경
- SkillListEditor로 이름을 변경
- SkillList라는 Molecule을 만들 것이다.
