+++ src/components/UI/organism/index.ts
@@ -4,7 +4,7 @@ export { default as SignupForm } from './SignupForm';
 export { default as AlertModal } from './AlertModal';
 export { default as ProfileInfoCard } from './ProfileInfoCard';
 export { default as ProfileIntroduce } from './ProfileIntroduce';
-export { default as SkillList } from './SkillList';
+export { default as SkillListEditor } from './SkillListEditor';
 export { default as AccountInfoModal } from './AccountInfoModal';
 export { default as Calendar } from './Calendar';
 export { default as ProjectModal } from './ProjectModal';
